+++ jdaviz/core/loaders/importers/catalog/__init__.py
@@ -1,1 +1,0 @@
-from .catalog import * # noqa