+++ jdaviz/core/loaders/importers/mos/mos.py
@@ -44,9 +44,9 @@ _MOS_PRODUCTS = {
               'viewer_reference': 'imviz-image-viewer',
               'viewer_traitlet_prefix': 'viewer_image'},
     'catalog': {'pattern': _CAT_PATTERN,
-                'format': 'Catalog',
-                'viewer_label': 'Table',
-                'viewer_reference': 'table-viewer',
+                'format': 'Source Catalog',
+                'viewer_label': 'Source Catalog Table',
+                'viewer_reference': 'source-catalog-table-viewer',
                 'viewer_traitlet_prefix': 'viewer_catalog'},
 }