+++ jdaviz/core/loaders/importers/source_catalog/test_load_source_catalogs.py
@@ -14,8 +14,8 @@ import pytest
 
 def _make_catalog(with_units=True, as_skycoord=False):
     """
-    Create a sample catalog table with optional units and SkyCoord columns
-    for testing purposes.
+    Create a sample source catalog table, with or without units and with or
+    without SkyCoord columns.
     """
 
     ra = [9.423, 9.421, 9.415] * (u.deg if with_units else 1)
@@ -33,6 +33,11 @@ def _make_catalog(with_units=True, as_skycoord=False):
 
 
 def _make_catalog_xy_radec(with_units=True, colnames=None):
+    """
+    Create a sample source catalog with both RA/Dec and X/Y source positions,
+    with or without units. Specify the column names using the `colnames` parameter
+    to override defaults.
+    """
 
     ra = [337.52274, 337.48273, 337.48296, 337.52333] * (u.deg if with_units else 1)
     dec = [-20.80742, -20.80741, -20.82380, -20.82425] * (u.deg if with_units else 1)
@@ -51,6 +56,10 @@ def _make_catalog_xy_radec(with_units=True, colnames=None):
 
 
 def _make_catalog_string_coord_columns():
+    """Create a sample source catalog with both RA/Dec and X/Y source positions,
+    using string representations for the coordinates.
+    """
+
     # complete nonsense coordinates, just care about parsing the units
     ra = ['5° 55′ 55″', '5° 55′ 55″', '5° 55′ 55″', '09h59m46.18778822s']
     dec = ['4° 44′ 44″', '4° 44′ 44″', '4° 44′ 44″', '+02d11m32.14597268s']
@@ -64,7 +73,7 @@ def _make_catalog_string_coord_columns():
 
 
 def _make_catalog_no_coordinates():
-    # Table to test loading catalogs that aren't 'Source Catalogs',
+    """Table to test loading catalogs that aren't 'Source Catalogs'."""
     col1 = ['A', 'B', 'C']
     col2 = [1, 2, 3]
     col3 = [True, False, True]
@@ -89,36 +98,35 @@ def test_load_catalog_no_source_positions(imviz_helper, image_2d_wcs):
 
     # load catalog, all columns
     imviz_helper.load(catalog_obj, col_other=['col1', 'col2', 'col3'],
-                      format='Catalog')
+                      format='Source Catalog')
 
     # check for the table in the data collection
     dc = imviz_helper._app.data_collection
     assert len(dc) == 2
-    assert 'Catalog' in imviz_helper._app.data_collection.labels
+    assert 'Source Catalog' in imviz_helper._app.data_collection.labels
     tab = imviz_helper._app.data_collection[1].get_object(Table)
     assert 'col1' in tab.colnames
 
     # make sure linking doesn't produce any errors when alignment changes.
-    # this isn't relevant for this catalog with no source positons,
-    # but orientation will check for the presence of certain
-    # components in a table to decide not to link and we
-    # want to make sure that works correctly
+    # this isn't relevant for this catalog with no source positons, but the
+    # orientation plugin will check for the presence of certain components in a
+    # table to decide not to link and we want to make this case is handled correctly.
     imviz_helper.plugins['Orientation'].align_by = 'WCS'
 
 
-def test_load_catalog_with_string_coord_cols(imviz_helper):
+def test_load_catalog_with_string_coord_cols(deconfigged_helper):
     """
     Test loading a catalog with string RA/Dec columns (that can be converted
-    into units, e.g string representation of hourangle units) into the
-    Imviz helper."""
+    into units, e.g string representation of hourangle units) into the app.
+    """
     catalog_obj = _make_catalog_string_coord_columns()
 
     # load catalog
-    imviz_helper.load(catalog_obj, format='Catalog')
+    deconfigged_helper.load(catalog_obj, format='Source Catalog')
 
-    dc = imviz_helper._app.data_collection
+    dc = deconfigged_helper._app.data_collection
     assert len(dc) == 1
-    assert 'Catalog' in imviz_helper._app.data_collection.labels
+    assert 'Source Catalog' in deconfigged_helper._app.data_collection.labels
 
     # make coordinate columns were renamed to Right Ascension and Declination,
     # X and Y in the data collection for consistency, and that RA / Dec always
@@ -147,8 +155,14 @@ def test_load_catalog_with_string_coord_cols(imviz_helper):
 
 @pytest.mark.parametrize("from_file", [True, False])
 @pytest.mark.parametrize("with_units", [True, False])
-def test_load_catalog_xy_and_radec(imviz_helper, tmp_path, from_file, with_units):
-    catalog_obj = _make_catalog_xy_radec(with_units=True)
+def test_load_catalog_xy_and_radec(deconfigged_helper, tmp_path, from_file, with_units):
+    """
+    Test loading a Source Catalog that has both X/Y and RA/Dec columns,
+    ensuring that the correct columns are recognized and loaded with appropriate
+    units (once from the original catalog, and once as assigned by the loader)
+    as well as testing reading from object and from file.
+    """
+    catalog_obj = _make_catalog_xy_radec(with_units)
 
     if from_file:
         fn = os.path.join(tmp_path, "catalog.ecsv")
@@ -158,16 +172,16 @@ def test_load_catalog_xy_and_radec(imviz_helper, tmp_path, from_file, with_units
         catalog = catalog_obj
 
     # load catalog
-    imviz_helper.load(catalog, format='Catalog')
+    deconfigged_helper.load(catalog, format='Source Catalog')
 
-    dc = imviz_helper._app.data_collection
+    dc = deconfigged_helper._app.data_collection
     assert len(dc) == 1
-    assert 'Catalog' in imviz_helper._app.data_collection.labels
+    assert 'Source Catalog' in deconfigged_helper._app.data_collection.labels
 
     # make sure 'RA' column was renamed to Right Ascension and 'Dec' to 'Declination'
     # in the data collection for consistency, and that the table in the data
     # collection always has units
-    qtab = imviz_helper._app.data_collection[0].get_object(QTable)
+    qtab = deconfigged_helper._app.data_collection[0].get_object(QTable)
     assert 'RA' in qtab.colnames
     assert 'Dec' in qtab.colnames
     assert 'X' in qtab.colnames
@@ -176,8 +190,10 @@ def test_load_catalog_xy_and_radec(imviz_helper, tmp_path, from_file, with_units
     # make sure only ra/dec/x/y/index/data link loaded, since we didn't specify more columns
     assert len(qtab.colnames) == 6
     # and that it has the correct contents
-    assert_quantity_allclose(qtab['RA'], catalog_obj['RA'])
-    assert_quantity_allclose(qtab['Dec'], catalog_obj['Dec'])
+    un = 1 if with_units else u.deg
+    assert_quantity_allclose(qtab['RA'], catalog_obj['RA'] * un)
+    assert_quantity_allclose(qtab['Dec'], catalog_obj['Dec'] * un)
+    assert qtab['RA'].unit == qtab['Dec'].unit == u.deg
     assert_quantity_allclose(qtab['X'], catalog_obj['X'])
     assert_quantity_allclose(qtab['Y'], catalog_obj['Y'])
 
@@ -196,7 +212,7 @@ def test_import_enabled_disabled(imviz_helper):
     ldr = loaders['object']
     ldr.object = catalog_obj
 
-    ldr.format = 'Catalog'
+    ldr.format = 'Source Catalog'
     ldr.importer.col_ra.selected = '---'
     ldr.importer.col_dec.selected = '---'
     ldr.importer.col_x.selected = '---'
@@ -249,12 +265,12 @@ def test_load_catalog(imviz_helper, image_2d_wcs, tmp_path, from_file, with_unit
     imviz_helper.plugins['Orientation'].align_by = 'WCS'
 
     # load catalog
-    imviz_helper.load(catalog, format='Catalog')
+    imviz_helper.load(catalog, format='Source Catalog')
 
-    # ensure that it is in the data collection with the correct label "Catalog"
+    # ensure that it is in the data collection with the correct label "Source Catalog"
     dc = imviz_helper._app.data_collection
     assert len(dc) == 3  # image, orientation layer, and catalog
-    assert 'Catalog' in imviz_helper._app.data_collection.labels
+    assert 'Source Catalog' in imviz_helper._app.data_collection.labels
 
     # make sure 'RA' column was renamed to Right Ascension and 'Dec' to 'Declination'
     # in the data collection for consistency, and that the table in the data
@@ -286,23 +302,23 @@ def test_load_catalog(imviz_helper, image_2d_wcs, tmp_path, from_file, with_unit
     # in the data collection
     ldr = imviz_helper.loaders['file' if from_file else 'object']
     setattr(ldr, 'filepath' if from_file else 'object', catalog)
-    ldr.format = 'Catalog'
+    ldr.format = 'Source Catalog'
     assert ldr.importer._obj.col_ra_has_unit == with_units
 
     # load it again, make sure label incremented by 1
-    imviz_helper.load(catalog, format='Catalog')
+    imviz_helper.load(catalog, format='Source Catalog')
     assert len(dc) == 4  # image, orientation layer, and 2 catalogs
-    assert 'Catalog (1)' in imviz_helper._app.data_collection.labels
+    assert 'Source Catalog (1)' in imviz_helper._app.data_collection.labels
 
     # load with custom label and check label
-    imviz_helper.load(catalog, data_label='my_catalog', format='Catalog')
+    imviz_helper.load(catalog, data_label='my_catalog', format='Source Catalog')
     assert len(dc) == 5  # image, orientation layer, and 3 catalogs
     assert 'my_catalog' in imviz_helper._app.data_collection.labels
 
     # test other loader API options. switch RA and Dec col just to test
     # non-default column selection
     imviz_helper.load(catalog, data_label='with_flux', col_other='flux',
-                      col_ra='Dec', col_dec='RA', format='Catalog')
+                      col_ra='Dec', col_dec='RA', format='Source Catalog')
     assert len(dc) == 6  # image, orientation layer, and 4 catalogs
     assert 'flux' in dc['with_flux'].get_object(QTable).colnames
     qtab = imviz_helper._app.data_collection[-1].get_object(QTable)
@@ -331,11 +347,11 @@ def test_load_catalog_skycoord(imviz_helper, tmp_path, from_file):
         catalog = catalog_obj
 
     # load catalog
-    imviz_helper.load(catalog, format='Catalog')
+    imviz_helper.load(catalog, format='Source Catalog')
 
     dc = imviz_helper._app.data_collection
     assert len(dc) == 1
-    assert 'Catalog' in imviz_helper._app.data_collection.labels
+    assert 'Source Catalog' in imviz_helper._app.data_collection.labels
 
     qtab = imviz_helper._app.data_collection[0].get_object(QTable)
     assert 'SkyCoord_RA' in qtab.colnames
@@ -370,9 +386,9 @@ def test_astroquery_load_catalog_source(deconfigged_helper):
     except requests.exceptions.RequestException as exc:
         pytest.skip(f"Transient remote archive failure: {exc}")
 
-    assert 'Catalog' in ldr.format.choices
+    assert 'Source Catalog' in ldr.format.choices
 
-    ldr.format = 'Catalog'
+    ldr.format = 'Source Catalog'
     ldr.importer.col_ra = 'ra'
     ldr.importer.col_dec = 'dec'
     ldr.importer.col_id = 'source_id'
@@ -381,7 +397,7 @@ def test_astroquery_load_catalog_source(deconfigged_helper):
     ldr.load()
 
     assert 'Scatter' in deconfigged_helper.viewers
-    assert 'Catalog' in deconfigged_helper.viewers['Scatter'].data_menu.layer.choices
+    assert 'Source Catalog' in deconfigged_helper.viewers['Scatter'].data_menu.layer.choices
 
 
 @pytest.mark.remote_data
@@ -425,8 +441,8 @@ def test_astroquery_load_catalog_from_viewer(deconfigged_helper):
         else:
             raise Exception(tb)
 
-    assert 'Catalog' in ldr.format.choices
-    ldr.format = 'Catalog'
+    assert 'Source Catalog' in ldr.format.choices
+    ldr.format = 'Source Catalog'
     ldr.load()
 
 
@@ -459,10 +475,10 @@ def test_astroquery_jwst_hst(deconfigged_helper, telescope):
     # note: querying coverage covered by test_resolver_table_as_query_astroquery
 
     ldr.treat_table_as_query = False
-    assert 'Catalog' in ldr.format.choices
+    assert 'Source Catalog' in ldr.format.choices
 
     # remove and replace with an assertion after JDAT-6412
-    ldr.format = 'Catalog'
+    ldr.format = 'Source Catalog'
 
     ldr.load()
     assert len(deconfigged_helper._app.data_collection) == 1
@@ -476,7 +492,7 @@ def test_invalid(imviz_helper, tmp_path):
     ldr = imviz_helper.loaders['object']
     with pytest.raises(ValueError, match="Parsed input is empty or None, cannot resolve."):
         ldr.object = empty_table
-    assert 'Catalog' not in ldr.format.choices
+    assert 'Source Catalog' not in ldr.format.choices
 
     # empty table from file
     fn = os.path.join(tmp_path, "empty_catalog.ecsv")
@@ -486,20 +502,20 @@ def test_invalid(imviz_helper, tmp_path):
     # Since no parsing of the actual file happens with the file parser
     # we don't error out as we did above.
     ldr.filepath = fn
-    assert 'Catalog' not in ldr.format.choices
+    assert 'Source Catalog' not in ldr.format.choices
 
     # while a fits file can be opened with table.read, it should not be
     # validated as a Catalog
     ldr = imviz_helper.loaders['object']
     ldr.object = fits.ImageHDU(np.ones((32, 25)))
     assert 'Image' in ldr.format.choices
-    assert 'Catalog' not in ldr.format.choices
+    assert 'Source Catalog' not in ldr.format.choices
 
 
 def test_scatter_viewer(deconfigged_helper):
     ldr = deconfigged_helper.loaders['object']
     ldr.object = _make_catalog(with_units=True)
-    ldr.format = 'Catalog'
+    ldr.format = 'Source Catalog'
 
     assert 'Scatter' in ldr.importer.viewer.create_new.choices
     ldr.importer.viewer.create_new = 'Scatter'
@@ -509,7 +525,7 @@ def test_scatter_viewer(deconfigged_helper):
     assert 'Scatter' in deconfigged_helper.new_viewers
 
     nv = deconfigged_helper.new_viewers['Scatter']
-    nv.dataset = ['Catalog']
+    nv.dataset = ['Source Catalog']
     nv.viewer_label = 'Added Scatter Viewer'
     nv()
 
@@ -519,7 +535,7 @@ def test_scatter_viewer(deconfigged_helper):
 def test_histogram_viewer(deconfigged_helper):
     ldr = deconfigged_helper.loaders['object']
     ldr.object = _make_catalog(with_units=True)
-    ldr.format = 'Catalog'
+    ldr.format = 'Source Catalog'
     ldr.importer.col_other = ['flux']
 
     assert 'Histogram' in ldr.importer.viewer.create_new.choices
@@ -530,7 +546,7 @@ def test_histogram_viewer(deconfigged_helper):
     assert 'Histogram' in deconfigged_helper.new_viewers
 
     nv = deconfigged_helper.new_viewers['Histogram']
-    nv.dataset = ['Catalog']
+    nv.dataset = ['Source Catalog']
     nv.xatt = 'flux'
     nv.viewer_label = 'Added Histogram Viewer'
     nv()
@@ -544,19 +560,22 @@ def test_histogram_viewer(deconfigged_helper):
     assert str(deconfigged_helper.viewers['Histogram']._obj.glue_viewer.state.x_att) == 'RA'  # noqa
 
 
-def test_table_viewer(deconfigged_helper):
+def test_source_catalog_table_viewer(deconfigged_helper):
+    """
+    Test basic functionality for the 'Source Catalog Table' viewer.
+    """
     ldr = deconfigged_helper.loaders['object']
     ldr.object = _make_catalog(with_units=True)
-    ldr.format = 'Catalog'
-    ldr.importer.viewer.create_new = 'Table'
+    ldr.format = 'Source Catalog'
+    ldr.importer.viewer.create_new = 'Source Catalog Table'
     ldr.load()
 
-    assert len(deconfigged_helper.viewers) == 1  # Table viewer created upon load
-    tv = deconfigged_helper.viewers['Table']
+    assert len(deconfigged_helper.viewers) == 1  # Source Catalog Table viewer created upon load
+    tv = deconfigged_helper.viewers['Source Catalog Table']
     assert len(tv._obj.glue_viewer.layers) == 1
 
     # create another viewer through viewer creator
-    vc = deconfigged_helper.new_viewers['Table']
+    vc = deconfigged_helper.new_viewers['Source Catalog Table']
     vc.dataset.select_all()
     nv = vc()
 
@@ -589,7 +608,8 @@ def test_load_catalog_from_hdulist(deconfigged_helper, tmp_path, from_file):
     """
     Test loading a catalog from a FITS file HDUList (from opened file in memory,
     and from a fits file) containing a BinTableHDU extension and verify that it
-    can be loaded through the catalog importer.
+    can be loaded through the Source Catalog importer into a Source Catalog Table
+    viewer.
     """
 
     # create an HDUList with a table extension
@@ -609,10 +629,10 @@ def test_load_catalog_from_hdulist(deconfigged_helper, tmp_path, from_file):
         ldr = deconfigged_helper.loaders['object']
         ldr.object = hdulist
 
-    # Verify Catalog format is available
-    assert 'Catalog' in ldr.format.choices
+    # Verify Source Catalog format is available
+    assert 'Source Catalog' in ldr.format.choices
 
-    ldr.format = 'Catalog'
+    ldr.format = 'Source Catalog'
 
     # Check that Primary HDU (index 0) is not in the extension choices, it should
     # have been filtered out because it is not a table extension
@@ -620,13 +640,18 @@ def test_load_catalog_from_hdulist(deconfigged_helper, tmp_path, from_file):
     assert len(extension_labels) == 1
     assert extension_labels[0] == '1: [CATALOG,1]'
 
-    # load catalog
+    # load catalog into a Source Catalog Table viewer
+    ldr.importer.viewer.create_new = 'Source Catalog Table'
     ldr.load()
 
     # verify catalog is in the data collection
     dc = deconfigged_helper._app.data_collection
     assert len(dc) == 1
-    assert 'Catalog' in dc.labels
+    assert 'Source Catalog' in dc.labels
+
+    # and was correctly loaded into a Source Catalog Table viewer
+    assert len(deconfigged_helper.viewers) == 1
+    assert 'Source Catalog Table' in deconfigged_helper.viewers
 
     # verify the table contents
     qtab = dc[0].get_object(QTable)
@@ -654,7 +679,7 @@ def test_catalog_visibility(imviz_helper, image_2d_wcs):
     table_x_y_only = orig_catalog['X', 'Y', 'Obj_ID']
 
     imviz_helper.load(table_ra_dec_only,
-                      data_label='catalog0', format='Catalog')
+                      data_label='catalog0', format='Source Catalog')
 
     # since we're pixel linked and catalog has only world coordinates,
     # visibility should be off by default
@@ -668,7 +693,7 @@ def test_catalog_visibility(imviz_helper, image_2d_wcs):
 
     # but if we load the catalog with X, Y, it should be visible
     imviz_helper.load(table_x_y_only,
-                      data_label='catalog1', format='Catalog')
+                      data_label='catalog1', format='Source Catalog')
     assert dm.data_labels_visible == ['catalog1', 'Image[DATA]']
 
     assert po.layer.choices == ['Image[DATA]', 'catalog1']  # catalog layer is now an option
@@ -679,7 +704,7 @@ def test_catalog_visibility(imviz_helper, image_2d_wcs):
     # load catalog with RA, Dec only again. Its default visibility should
     # now be on since we're WCS linked
     imviz_helper.load(table_ra_dec_only,
-                      data_label='catalog2', format='Catalog')
+                      data_label='catalog2', format='Source Catalog')
 
     # the pixel-only 'catalog1' should now be hidden
     assert dm.data_labels_visible == ['catalog2', 'Image[DATA]']
@@ -687,7 +712,7 @@ def test_catalog_visibility(imviz_helper, image_2d_wcs):
     # loading a pixel-coordinate-only catalog should now be hidden by default
     # since were WCS linked
     imviz_helper.load(table_x_y_only,
-                      data_label='catalog3', format='Catalog')
+                      data_label='catalog3', format='Source Catalog')
     assert 'catalog3' not in dm.data_labels_visible
 
 
@@ -718,13 +743,14 @@ def test_hdulist_multiple_table_extensions(deconfigged_helper):
     ldr = deconfigged_helper.loaders['object']
     ldr.object = hdulist
 
-    # This HDUList contains both table and image extensions, so both Catalog and Image
-    # formats should be valid. Verify Catalog is listed last when multiple formats exist.
-    assert 'Catalog' in ldr.format.choices
+    # This HDUList contains both table and image extensions, so both
+    # Source Catalog and Image formats should be valid. Verify Catalog is listed
+    # last when multiple formats exist.
+    assert 'Source Catalog' in ldr.format.choices
     assert len(ldr.format.choices) > 1
-    assert ldr.format.choices[-1] == 'Catalog'
+    assert ldr.format.choices[-1] == 'Source Catalog'
 
-    ldr.format = 'Catalog'
+    ldr.format = 'Source Catalog'
 
     # check that both table extensions are available. Checking this indirectly
     # verifies that the primary and image HDUs were correctly filtered out of
@@ -806,7 +832,7 @@ def test_load_catalog_from_fits_multiselect(deconfigged_helper):
 
     ldr = deconfigged_helper.loaders['object']
     ldr.object = hdul
-    ldr.format = 'Catalog'
+    ldr.format = 'Source Catalog'
 
     assert ldr.importer.extension.multiselect is True
 
@@ -854,7 +880,7 @@ def test_load_catalog_from_fits_multiselect(deconfigged_helper):
                           fits.BinTableHDU(table3)])
 
     ldr.object = hdul2
-    ldr.format = 'Catalog'
+    ldr.format = 'Source Catalog'
 
     ldr.importer.extension.selected = ldr.importer.extension.choices