+++ jdaviz/core/template_mixin.py
@@ -2454,7 +2454,7 @@ class LayerSelect(SelectPluginComponent):
                 return True
 
             # non-catalog layers should remain available regardless of link type
-            if getattr(lyr, 'meta', {}).get('_importer', '') != 'CatalogImporter':
+            if getattr(lyr, 'meta', {}).get('_importer', '') != 'SourceCatalogImporter':
                 return True
 
             comp_labels = [str(x) for x in lyr.component_ids()]
@@ -4933,9 +4933,17 @@ class DatasetSelect(SelectPluginComponent):
         def is_image(data):
             return len(data.shape) == 2
 
+        def is_source_catalog_table(data):
+            return data.meta.get('_importer', '') == 'SourceCatalogImporter'
+
+        def is_spectral_lines_list_table(data):
+            return data.meta.get('_importer', '') == 'SpectralLinesImporter'
+
+        def is_generic_table(data):
+            return data.meta.get('_importer', '') == 'GenericCatalogImporter'
+
         def is_catalog(data):
-            return data.meta.get('_importer', '') in ['SpectralLinesImporter',
-                                                      'CatalogImporter']
+            return is_source_catalog_table(data) or is_spectral_lines_list_table(data) or is_generic_table(data)  # noqa
 
         def is_catalog_or_image_not_spectrum(data):
             return is_catalog(data) or is_image_not_spectrum(data)