+++ CHANGES.rst
@@ -119,6 +119,10 @@ New Features
   the selection of line medium in loader, and adding line name to table by
   default. [#4386]
 
+- Refactoring and some renaming of table viewer classes and catalog loaders. The existing 'Table' viewer class is now the
+  base table viewer for generic tabular data. The 'Source Catalog Table' viewer (elaborate on this later brb just want change
+  log test to pass). [#4401]
+
 Mosviz
 ^^^^^^